+++ samples/extensions/failure-notify/failure-notify.h
@@ -8,7 +8,7 @@ using namespace residue;
 class FailureNotify : public DispatchErrorExtension {
 public:
 
-	FailureNotify() : DispatchErrorExtension("failure-notify"), m_failureCount(0UL) 
+	FailureNotify() : DispatchErrorExtension("failure-notify"), m_failureCount(0UL), m_sent(false)
 	{
 		
 	}
@@ -16,8 +16,10 @@ public:
 	virtual ~FailureNotify() = default;
 
 	virtual Extension::Result execute(const DispatchErrorExtension::Data* const) override;
+	virtual void reset() override;
 private:
 	unsigned long m_failureCount;
+        bool m_sent;
 
 	void notifyRecipients(const DispatchErrorExtension::Data* const data);
 };